Poc Analisis de sentimientos/Cod/requirements.txt
absl-py==0.9.0
altgraph==0.17
argcomplete==1.10.0
astroid==2.4.2
astunparse==1.6.3
azure-ai-textanalytics==1.0.0
azure-common==1.1.25
azure-core==1.7.0
backcall==0.2.0
beautifulsoup4==4.8.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
distro==1.5.0
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
future==0.18.2
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
imageio==2.8.0
IMAPClient==2.1.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
jedi==0.17.1
joblib==0.15.1
jupyter-client==6.1.5
jupyter-core==4.6.3
Keras==2.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.1
Markdown==3.2.2
matplotlib==3.2.2
mccabe==0.6.1
msrest==0.6.17
networkx==2.4
nltk==3.5
numpy==1.19.0
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.2.1
pandas==1.0.5
parso==0.7.0
pdfminer==20191125
pdfminer.six==20181108
pdftotree==0.4.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
protobuf==3.12.2
py2exe==0.9.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
Pygments==2.6.1
PyInstaller==3.6
pylint==2.5.3
pyparsing==2.4.7
PyPDF2==1.26.0
PySocks==1.7.1
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
selectivesearch==0.4
six==1.12.0
sklearn==0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SpeechRecognition==3.8.1
tabula-py==2.1.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
textblob==0.15.3
textract==1.6.3
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
tweepy==3.9.0
tzlocal==1.5.1
urllib3==1.25.9
vaderSentiment==3.3.2
Wand==0.6.1
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.9
